fix(stocking): move ranges in their package
